move extractors/ida/__init__.py logic to extractor.py
